+++ app/models/OtherArticle.java
@@ -5,7 +5,7 @@ import java.sql.Date;
 
 public class OtherArticle extends Article {
 
-    public OtherArticle(String name, Integer price, Integer condition, String description, Date creationDate, String image, String type) {
+    public OtherArticle(String name, Double price, Integer condition, String description, Date creationDate, String image, String type) {
         super(name, price, condition, description, creationDate, image, type);
     }
 }
